move Plugin from plugins/__init__.py to plugins/contracts.py
Clearer file name, less code in __init__.py (didn't really need to be
there), room for additional contracts.
